Validation moved to views.py (ValidationView)
